Add option to disable async
This is done by checking if the source contains the word 'async'.
This catches await, since it can only be used in async functions.
It wraps eval and function ctors and injects the async check there too.
This has some side effects like no direct eval support.
The option is currently opt in.